+++ packages/teams_memory/CHANGELOG.md
@@ -7,11 +7,13 @@
 - Added method `BaseMemoryModule.get_memories_with_attributions` to easily get the messages associated to memories.
 - Turn on mypy strict mode and fix mypy errors
 - Added method `BaseMemoryModule.ask` to answer a question using memories.
+- Added support for not requiring a query or topic to search memories. Previously, you had to provide one of these.
 
 ### Bug Fixes
 
 - Improved documentation for sample and teams-memory integration.
 - Update the way we build the litellm client for better typings
+- Turn on mypy strict mode and fix mypy errors
 
 ## 0.1.2-alpha